feat: move infmon-config source files into src/common/src/config/
- Move crud.rs, model.rs, parse.rs, validate.rs via git mv
- Move lib.rs to mod.rs (preserving module declarations and pub use)
- Update internal crate:: references to crate::config::
- Add re-export shim in infmon-config for backward compatibility

Closes #100
